Allow Spout to be run from an IDE
This commit allows Spout to be run from an IDE. It has been tested with eclipse only so far.
This adds the possibility to load plugins directly from the class path instead of jars.
Only the server has been tested so far, client will probably be somewhat harder to do.
Debugging within an IDE has several advantages:

* Full breakpoint support without using the way slower remote debugging
* Class hot swapping without needing additional software like JRebel
  (using JRebel with an IDE should still be possible though)
* Faster development cycle: no need to rebuild all the jars just to make a quick test
* Easier navigation from stack traces to classes
* and many more…

A description on how to setup eclipse for this workflow can be found at: <add gist here later>

Signed-off-by: fr1zle <f@br0tbox.de>